+++ backend/api/serializers.py
@@ -5,7 +5,7 @@ from djoser.serializers import UserCreateSerializer, UserSerializer
 from drf_extra_fields.fields import Base64ImageField
 from rest_framework import serializers
 from rest_framework.fields import SerializerMethodField
-
+from api.serializers.user_serializers import UserProfileSerializer
 from job.models import CaseImage, Comment, FavoriteOrder, Sphere, Case, Favorite, Instrument, Skill
 from users.models import Subscription
 
@@ -110,45 +110,6 @@ class CaseCreateSerializer(serializers.ModelSerializer):
         return CaseSerializer(instance).data
 
 
-class UserProfileSerializer(UserSerializer):
-    """
-    Сериализатор профиля пользователя.
-
-    Атрибуты:
-        is_subscribed (SerializerMethodField): поле, указывающее,
-        подписан ли текущий пользователь на автора.
-
-    Методы:
-        get_is_subscribed(obj: User) -> bool: возвращает True,
-        если текущий пользователь подписан на автора, иначе False.
-
-    """
-
-    is_subscribed = SerializerMethodField(read_only=True)
-
-    class Meta:
-        model = User
-        fields = ('email', 'id', 'username', 'first_name',
-                  'last_name', 'is_subscribed')
-
-    def get_is_subscribed(self, obj: User) -> bool:
-        user = self.context.get('request').user
-        if user.is_anonymous:
-            return False
-        return user.subscriber.filter(author=obj).exists()
-
-
-class UserProfileCreateSerializer(UserCreateSerializer):
-    """
-    Сериализатор для создания пользователя.
-
-    """
-
-    class Meta:
-        model = User
-        fields = ('email', 'username', 'first_name', 'last_name', 'password')
-
-
 class SubscriptionSerializer(UserProfileSerializer):
     """
     Сериализатор для подписки на автора.
@@ -187,15 +148,6 @@ class SubscriptionSerializer(UserProfileSerializer):
                 {'subscription': ['Нельзя подписаться на себя']}
             )
         return data
-
-
-# заглушка
-class CaseSerializer():
-    pass
-
-# заглушка
-class OrderSerializer():
-    pass
 
 
 class CaseImageSerializer(serializers.ModelSerializer):
